+++ src/main/java/com/facilio/bmsconsole/actions/WorkorderToolsAction.java
@@ -77,7 +77,7 @@ public class WorkorderToolsAction extends FacilioAction {
 		long qty = (long) context.get(FacilioConstants.ContextNames.TOTAL_QUANTITY);
 		setResult(FacilioConstants.ContextNames.TOTAL_QUANTITY, qty);
 		setResult(FacilioConstants.ContextNames.TOTAL_COST, cost);
-		long totalCost = (long) context.get(FacilioConstants.ContextNames.WO_TOTAL_COST);
+		double totalCost = (double) context.get(FacilioConstants.ContextNames.WO_TOTAL_COST);
 		setResult(FacilioConstants.ContextNames.WO_TOTAL_COST, totalCost);
 		setResult("workorderToolsIds", workorderToolsIds);
 		return SUCCESS;
@@ -99,7 +99,7 @@ public class WorkorderToolsAction extends FacilioAction {
 		long qty = (long) context.get(FacilioConstants.ContextNames.TOTAL_QUANTITY);
 		setResult(FacilioConstants.ContextNames.TOTAL_QUANTITY, qty);
 		setResult(FacilioConstants.ContextNames.TOTAL_COST, cost);
-		long totalCost = (long) context.get(FacilioConstants.ContextNames.WO_TOTAL_COST);
+		double totalCost = (double) context.get(FacilioConstants.ContextNames.WO_TOTAL_COST);
 		setResult(FacilioConstants.ContextNames.WO_TOTAL_COST, totalCost);
 		setResult("workorderToolsIds", workorderToolsIds);
 		return SUCCESS;
